Added audit trailing for category and product deletion

+++ ECommerce.Ui/Services/CategoryService.cs
@@ -80,9 +80,9 @@ namespace ECommerce.Ui.Services
             return success;
         }
 
-        public async Task<Boolean> Delete(long id)
+        public async Task<Boolean> Delete(long id, string userId)
         {
-            var response = await _httpClient.DeleteAsync($"{_route}/{id}");
+            var response = await _httpClient.DeleteAsync($"{_route}/{userId}/{id}");
             
             if (response.StatusCode == System.Net.HttpStatusCode.BadRequest)
             {
